third_party/blink/web_tests/external/wpt/css/css-shadow-parts/host-part-001-expected.txt

This is a testharness.js-based test.
FAIL :host::part works assert_equals: :host::part() works expected "rgb(0, 255, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
